# improvement: Use PcCollector for SemanticTokenProvider