--- a/tests/unit/src/test/resources/semanticTokens/example/TryCatch.scala
+++ b/tests/unit/src/test/resources/semanticTokens/example/TryCatch.scala
@@ -5,7 +5,7 @@
     <<val>>/*keyword*/ <<x>>/*variable,readonly*/ = <<2>>/*number*/
     <<x>>/*variable,readonly*/ <<+>>/*method,abstract*/ <<2>>/*number*/
   } <<catch>>/*keyword*/ {
-    <<case>>/*keyword*/ t: <<Throwable>>/*type*/ <<=>>>/*operator*/
+    <<case>>/*keyword*/ <<t>>/*variable,readonly*/: <<Throwable>>/*type*/ <<=>>>/*operator*/
       <<t>>/*variable,readonly*/.<<printStackTrace>>/*method*/()
   } <<finally>>/*keyword*/ {
     <<val>>/*keyword*/ <<text>>/*variable,readonly*/ = <<"">>/*string*/