… tests/slow/src/test/scala/tests/sbt/SbtServerSuite.scala
… tests/unit/src/main/scala/tests/BaseWorksheetLspSuite.scala
… tests/unit/src/test/resources/semanticTokens/example/MacroAnnotation.scala