Previously `SemanticTokenProvider` had its own tree traverser for collecting all symbol occurrences.
This PR changes it to use `PcCollector`, which should make optimisation and bugfixes easier. It should also simplify adding `SemanticTokens` for Scala 3