@jkciesluk 2023-01-17
Quick benchmarks show that after these changes, semantic tokens can work slower
After chages:
```
[info] Benchmark                                 (currentHighlightRequest)  Mode  Cnt    Score    Error  Units
[info] SemanticHighlightBench.semanticHighlight                    A.scala    ss   10    2.641 ±  0.508  ms/op
[info] SemanticHighlightBench.semanticHighlight        ClusterDaemon.scala    ss   10   49.994 ±  6.692  ms/op
[info] SemanticHighlightBench.semanticHighlight               Typers.scala    ss   10  161.479 ± 30.930  ms/op
[info] SemanticHighlightBench.semanticHighlight                Exprs.scala    ss   10   25.018 ± 14.654  ms/op
```
Before change:
```
[info] Benchmark                                 (currentHighlightRequest)  Mode  Cnt    Score    Error  Units
[info] SemanticHighlightBench.semanticHighlight                    A.scala    ss   10    1.912 ±  0.498  ms/op
[info] SemanticHighlightBench.semanticHighlight        ClusterDaemon.scala    ss   10   47.090 ± 16.405  ms/op
[info] SemanticHighlightBench.semanticHighlight               Typers.scala    ss   10  127.240 ± 45.553  ms/op
[info] SemanticHighlightBench.semanticHighlight                Exprs.scala    ss   10   22.164 ± 17.558  ms/op
```

On the other hand, they are more correct now and we can optimise `PcCollector`, which will improve also `highlight` and `rename`